# api/openapi.yaml
openapi: 3.0.3
info:
  title: Self-Hosted RWA Platform API
  version: 1.0.0
  description: >
    FROZEN API skeleton (lead-owned). State-changing endpoints accept an
    Idempotency-Key header. Errors use stable machine-readable codes.
    On-chain actions (investor buy/request/claim/cancel, admin/treasury/redemption
    wallet actions) are encoded client-side by the SPA and broadcast from the
    connected wallet — never executed by a custodial server key.
servers:
  - url: http://localhost:8080
security:
  - BearerJwt: []
...
paths:
...
  /api/v1/me/wallet-status:
    get:
      tags: [compliance]
      operationId: getMyWalletStatus
      summary: This session's own wallet status (subject-scoped).
      description: >
        Returns the WalletStatus for the address bound to the presented wallet session
        only. Authenticated by the X-Wallet-Session bearer minted at challenge-verify,
        NOT by an operator API key. Replaces the investor page's former dependency on
        the operator-only global wallet list.
      security:
        - WalletSession: []
      responses:
        "200": { description: Own wallet status, content: { application/json: { schema: { $ref: "#/components/schemas/WalletStatus" } } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/compliance/allowed/{address}:
    get:
      tags: [compliance]
      operationId: isAddressAllowed
      summary: Public transfer-preflight eligibility.
      description: >
        Minimal, unauthenticated lookup for transfer-recipient preflight. Returns only
        whether a transfer to `address` would currently be permitted — deliberately not
        the full WalletStatus, so an anonymous caller cannot enumerate a third party's
        exact status / validUntil / ownership.
      security: []
      parameters:
        - { name: address, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Eligibility, content: { application/json: { schema: { $ref: "#/components/schemas/AllowedResult" } } } }
...
components:
  securitySchemes:
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Admin JWT (HMAC-signed) issued by POST /auth/session after a successful
        wallet-signature challenge whose signer equals the configured project admin
        address. Attach as `Authorization: Bearer <jwt>`. Stateless — no server-side
        revocation; logout drops the token client-side. Investor read routes use
        X-Wallet-Session instead; requests with no/invalid token fall through to the
        public (read-only) surface.
    WalletSession:
      type: apiKey
      in: header
      name: X-Wallet-Session
      description: >
        Short-lived (≈15 min) opaque session token minted at challenge-verify, scoped to
        a single wallet address. Grants read access to that address's own status only —
        never an admin capability. Kept separate from the admin BearerJwt so the two
        auth models never collide.
...
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
...
    WalletStatus:
      type: object
      properties:
        address: { type: string }
        status: { type: string, enum: [Unknown, Allowed, Blocked] }
        validUntil: { type: integer }
        ownershipVerified: { type: boolean }
...
    AllowedResult:
      type: object
      required: [allowed]
      properties:
        allowed: { type: boolean, description: "Whether a transfer to the address is currently permitted. Nothing else is disclosed." }
...